GUI/rov-dashboard/requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
numpy>=1.24.3
pillow>=10.0.0
matplotlib>=3.7.1
pyserial>=3.5